makefile???????

+++ libft/ft_pow.c
@@ -6,13 +6,13 @@
 /*   By: emaveric <emaveric@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/17 17:48:12 by emaveric          #+#    #+#             */
-/*   Updated: 2020/02/19 17:48:12 by emaveric         ###   ########.fr       */
+/*   Updated: 2020/02/22 13:41:42 by emaveric         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-ULL			ft_pow(ULL n, ULL pow)
+unsigned long long	ft_pow(unsigned long long n, unsigned long long pow)
 {
 	return (pow ? n * ft_pow(n, pow - 1) : 1);
 }
